Refactor user token and other utils into a separate module
Also makes verify_user_token return the user's username using the token
